requirements.txt
# Telegram 客户端
telethon==1.34.0

# 异步 HTTP 客户端
aiohttp==3.9.1

# 时区支持
pytz==2024.1

# 环境变量管理
python-dotenv==1.0.0

# SOCKS 代理支持（可选，用于生成 Session）
python-socks[asyncio]==2.4.3
